requirements.txt
bleach
functions-framework
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-secret-manager
google-genai
jinja2
psutil
pydantic
python-dotenv
structlog
